src/rb-facematch/pyproject.toml
[tool.poetry]
name = "rb-facematch"
version = "0.1.0"
description = ""
authors = ["Rescue Lab <sahilsharma@umass.edu>"]
packages = [{include = "rb_facematch"}]

[tool.poetry.scripts]
rb_facematch = "rb_facematch.main:app"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
deepface = "*"
#tensorflow = "*" does not work on windows
tensorflow = {version = "*", markers = 'sys_platform != "win32"' }
tensorflow-intel = {version = "*", markers = 'sys_platform == "win32"' }
black = "*"
isort = "*"
tf-keras = "*"
jax = "*"
ultralytics = "*"
pandas = "*"
# 1.2.2 does not work
scikit-learn = "*"
ruff = "*"
cx-freeze = "*"
faiss-cpu = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
